Add tests for login/logout functionality
Previously, there were no tests for the login/logout functionality. This commit adds those tests by separating out the logic of the doGet() function in the /login servlet into another function getAuthentication(). This getAuthentication() function is tested via JUnit for correctness by creating an expected return in the LoginTest.java file.

For clarification, the Login.java and Authentication.java files are moved into sps/login so that they are in the same package and can be easily accessed by the JUnit tests in LoginTest.java. Additionally, a number of necessary dependencies for appengine testing are added to the pom.xml file.