Add register page and servlet with tests.
Previously, there was only a login/logout functionality. This commit adds a page where users can register for an account with their name and email which is stored in a datastore through a servlet. Regarding testing of the servlet, there are JUnit and Mockito tests. Regarding testing of the DOM manipulation in JavaScript, this commit also adds NPM and Node through Maven Frontend Plugin in order to use Jest and Jest Fetch Mock.