Add bootstrap to frontend
Previously, we used completely vanilla HTML, CSS. This commit adds bootstrap to the frontend, preserving previous functionality, general structure, and naming schemes except for switching the header with a navbar and putting the text into a card. Additionally, based on recommendations by UX designers and feedback from user testing, it adds a color gradient background to emphasize the text on the card as well as two buttons (one for an associated link and one for completing the deed).