Replace strings with constants and hide API keys.
Previously, there were many strings throughout the javascript code and the API key was shown in the script-auth.js file. This commit replaces the strings with constants at the beginning of the file that define the string as well as hides the API key in the config.js file and prevents it from being committed by adding it to .gitignore. It also acknowledges that making testing more convenient by testing the direct functions throughout the login javascript files is being sacrificed for more functionality.