Remove did deed button and reload after profle update
Previously, there was a did deed button on the homepage and after profile updates, there would be no reload. Because the did deed functionality was unable to be implemented in the timeframe and the lack of reload makes it unclear whether the profile was updated, this commit removes the did deed button and reloads the page after a profile update.